Football_Database/API/requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.95.1
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2==2.9.6
pydantic==1.10.7
python-dateutil==2.8.2
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.26.1
strawberry-graphql==0.175.0
typing_extensions==4.5.0
uvicorn==0.22.0
